+++ backend/views/partials/foot.ejs
